Break ConversationService into the five services it already was
1495 lines and 36 methods covering conversation storage, turn lifecycle,
approval resolution, pause queries and authorization. Five services in a
trenchcoat, and the seams were already visible: `_build_resume_tool_return` had
125 lines and touched nothing the class held except a unit of work.

The public class and every public signature stay exactly as they were. That is
not caution, it is a hard constraint: `agent_surfaces/services/
pending_interaction_resume.py` reaches `get_pending_user_interaction` and
`resolve_user_approval_internal` off one import of `ConversationService`. Split
those onto two public classes and that file needs two imports, `agent_surfaces
->agent` goes 6 -> 7, and CI fails in a module this change never touches.

Five collaborators, each with its dependencies stated in a constructor:

* `conversation_queries` -- the read half. Every method is resolve-agent,
  load-row, check-access, then read; they are together because that preamble is
  the part worth getting right.
* `conversation_turns` -- start and stop, the half that needs the conversation
  lock and the outbox.
* `conversation_approvals` -- resolving a pause. Three things are idempotent
  there for three different reasons, which is now written down.
* `conversation_resume_return` -- building the tool return an approval unblocks,
  including the principal switch that runs the approved tool as the user.
* `conversation_access` -- gained the permission checks, because every one of
  the above needed them.

`PauseResume` stopped being a mixin along the way. Its own docstring said it
lived in its own module "because that class is already well past the ratchet's
limit" -- relocation, not decomposition -- and it depended on a private method
of its host, which is exactly the bug that shape invites: this change deleted
that method and nothing complained until a test ran.

Three tests moved off a `__new__`-bypassed `ConversationService` and onto the
real collaborator they were always testing. `_pending_user_approval_from_messages`
had no callers at all and is gone.
