+++ lemma-backend/app/modules/agent/services/conversation_service.py
@@ -2,35 +2,35 @@
 
 from __future__ import annotations
 
-from collections.abc import Sequence
-from functools import partial
 from uuid import UUID
 
 
-from app.modules.agent.services.run_dispatch import run_enqueue_suppressed
+from app.modules.agent.services.conversation_approvals import (
+    ApprovalCoordinator,
+)
+from app.modules.agent.services.conversation_queries import ConversationQueries
+from app.modules.agent.services.conversation_turns import TurnCoordinator
+from app.modules.agent.services.conversation_resume_return import (
+    ResumeToolReturnBuilder,
+)
 from app.modules.agent.domain.sentinels import UNSET, UnsetType
-from app.core.authorization.context import ResourceRef, ResourceType
-from app.core.authorization.current import get_current_context
 from app.core.authorization.permissions import Permissions
 from app.core.infrastructure.db.uow import SqlAlchemyUnitOfWork
 from app.modules.agent.services.conversation_access import (
-    resolve_agent,
+    authorized_conversation,
+    resolve_expected_agent_id,
+    require_agent_action,
+    require_agent_actions,
+    resolve_agent_for_path,
     validate_conversation_access,
 )
 from app.modules.agent.domain.entities import (
-    Agent,
     AgentRun,
     Conversation,
     Message,
 )
 from app.modules.agent.domain.errors import (
-    AgentNotFoundError,
     ConversationNotFoundError,
-    UnknownApprovalError,
-)
-from app.modules.agent.domain.events import (
-    AgentRunStartedEvent,
-    AgentRunStopRequestedEvent,
 )
 from app.modules.agent.domain.ports import (
     AgentRepository,
@@ -40,53 +40,22 @@ from app.modules.agent.domain.approvals import ApprovalResolution
 from app.modules.agent.domain.value_objects import (
     AgentRunApprovalDecision,
     AgentRunStartResult,
-    AgentRunStatus,
     AgentRuntimeConfig,
     ConversationAgentSelection,
     ConversationStatus,
     ConversationType,
-    MessageDraft,
-    MessageKind,
-    MessageRole,
 )
-from app.modules.agent.domain.agent_host_permissions import (
-    agent_host_permission_request,
-)
-from app.modules.agent.services.approval_reconciliation import (
-    agent_host_permission_tool_return,
-    execute_approved_tool_as_user,
-    pending_user_approval_messages,
-    queue_approval_reconciliation,
-    record_session_approvals,
-    should_defer_approved_tool,
-)
-from app.modules.agent.services.runtime_profile_service import (
-    DEFAULT_SYSTEM_AGENT_RUNTIME_PROFILE_ID,
-)
-from app.modules.agent.services.realtime import (
-    input_added_payload,
-    message_payload,
-    publish_conversation_event,
-)
-from app.modules.agent.services.serialization import message_to_payload
 from app.modules.agent.services.workspace_location import (
     apply_location_metadata,
-    resolve_workspace_location,
 )
-from app.modules.pod.contracts import PodConfig
 from app.composition.agent_pod import create_agent_pod_repository
-from app.composition.agent_usage import UsageLimitExceededError, UsageService
-from app.composition.agent_snooze_scheduler import cancel_snooze_wake
+from app.composition.agent_usage import UsageService
 from app.modules.agent.infrastructure.wait_repository import (
     AgentConversationWaitRepository,
 )
 from app.modules.agent.services.pause_resume import (
     PAUSING_TOOL_NAMES,
-    PauseResumeMixin,
-)
-from app.modules.agent.tools.snooze.models import (
-    build_snooze_result,
-    elapsed_seconds,
+    PauseResume,
 )
 
 
@@ -95,7 +64,7 @@ from app.modules.agent.tools.snooze.models import (
 _PAUSING_TOOL_NAMES = PAUSING_TOOL_NAMES
 
 
-class ConversationService(PauseResumeMixin):
+class ConversationService:
     """Application service for conversation storage and run coordination."""
 
     def __init__(
@@ -114,6 +83,22 @@ class ConversationService(PauseResumeMixin):
         self.authorization_service = authorization_service
         self.fallback_model_name = fallback_model_name
         self.usage_service = usage_service
+        self.resume_returns = ResumeToolReturnBuilder(uow, agent_repository)
+        self.pauses = PauseResume(uow, conversation_repository, agent_repository)
+        self.approvals = ApprovalCoordinator(
+            uow, conversation_repository, self.resume_returns, self.pauses
+        )
+        self.queries = ConversationQueries(
+            uow, conversation_repository, agent_repository
+        )
+        self.turns = TurnCoordinator(
+            uow,
+            conversation_repository,
+            agent_repository,
+            self.approvals,
+            self.pauses,
+            usage_service,
+        )
 
     async def create_conversation(
         self,
@@ -131,7 +116,9 @@ class ConversationService(PauseResumeMixin):
     ) -> Conversation:
         organization_id = await self._get_pod_organization_id(pod_id)
         agent = (
-            await self._resolve_agent_for_path(pod_id=pod_id, agent_name=agent_name)
+            await resolve_agent_for_path(
+                self.agent_repository, pod_id=pod_id, agent_name=agent_name
+            )
             if agent_name is not None
             else None
         )
@@ -149,7 +136,7 @@ class ConversationService(PauseResumeMixin):
             actions = [Permissions.AGENT_EXECUTE]
             if agent is not None:
                 actions.append(Permissions.AGENT_READ)
-            await self._require_agent_actions(
+            await require_agent_actions(
                 user_id=user_id,
                 pod_id=pod_id,
                 agent_id=agent.id if agent else None,
@@ -200,95 +187,6 @@ class ConversationService(PauseResumeMixin):
 
         await apply_location_metadata(conversation, fetch_parent=_parent)
 
-    async def list_conversations(
-        self,
-        *,
-        pod_id: UUID,
-        agent_selection: ConversationAgentSelection[str],
-        user_id: UUID,
-        status: ConversationStatus | None = None,
-        type: ConversationType | None = None,
-        metadata_filters: dict[str, object] | None = None,
-        parent_id: UUID | None = None,
-        cursor: UUID | None = None,
-        limit: int = 20,
-    ) -> tuple[list[Conversation], UUID | None]:
-        expected_agent_id = await self._expected_agent_id(
-            pod_id=pod_id,
-            agent_name=agent_selection.value,
-        )
-        resolved_selection = agent_selection.resolve(expected_agent_id)
-        await self._require_agent_action(
-            user_id=user_id,
-            pod_id=pod_id,
-            agent_id=expected_agent_id,
-            action=Permissions.AGENT_READ,
-        )
-        return await self.conversation_repository.list_conversations(
-            user_id=user_id,
-            pod_id=pod_id,
-            agent_selection=resolved_selection,
-            status=status,
-            conversation_type=type,
-            metadata_filters=metadata_filters,
-            parent_id=parent_id,
-            cursor=cursor,
-            limit=limit,
-        )
-
-    async def get_conversation(
-        self,
-        *,
-        conversation_id: UUID,
-        user_id: UUID,
-        pod_id: UUID,
-        agent_name: str | None = None,
-        require_read_grant: bool = True,
-    ) -> Conversation:
-        expected_agent_id = await self._expected_agent_id(
-            pod_id=pod_id,
-            agent_name=agent_name,
-        )
-        # The latest run carries the failure diagnostics and the retry decision.
-        conversation = await self.conversation_repository.get_conversation(
-            conversation_id,
-            include_runs=True,
-        )
-        validate_conversation_access(
-            conversation,
-            user_id=user_id,
-            pod_id=pod_id,
-            agent_id=expected_agent_id,
-        )
-        # require_read_grant is False only for self-spawn (an agent operating on
-        # its own conversation tree): ownership is validated above, but the agent
-        # holds no agent.read grant on itself, so skip the cross-agent grant check.
-        if require_read_grant:
-            await self._require_agent_action(
-                user_id=user_id,
-                pod_id=pod_id,
-                agent_id=conversation.agent_id,
-                action=Permissions.AGENT_READ,
-            )
-        conversation.last_run_retryable = await self._latest_run_is_retryable(
-            conversation
-        )
-        return conversation
-
-    async def _latest_run_is_retryable(self, conversation: Conversation) -> bool:
-        """Whether the newest run can be replayed without duplicating output.
-
-        The status check runs first and short-circuits: a run that did not fail
-        is not retryable whatever its messages say, and that is nearly every
-        conversation, so the message query below is rarely reached. Asking the
-        database that question directly is what replaced eager-loading the
-        whole transcript to evaluate it in Python.
-        """
-        latest = conversation.agent_runs[-1] if conversation.agent_runs else None
-        if latest is None or latest.status != AgentRunStatus.FAILED:
-            return False
-        return await self.conversation_repository.run_has_only_user_messages(latest.id)
-
     async def update_conversation(
         self,
         *,
@@ -301,7 +199,8 @@ class ConversationService(PauseResumeMixin):
         agent_runtime: AgentRuntimeConfig | None | UnsetType = UNSET,
         metadata: dict[str, object] | None | UnsetType = UNSET,
     ) -> Conversation:
-        expected_agent_id = await self._expected_agent_id(
+        expected_agent_id = await resolve_expected_agent_id(
+            self.agent_repository,
             pod_id=pod_id,
             agent_name=agent_name,
         )
@@ -316,7 +215,7 @@ class ConversationService(PauseResumeMixin):
         )
         if conversation is None:
             raise ConversationNotFoundError()
-        await self._require_agent_action(
+        await require_agent_action(
             user_id=user_id,
             pod_id=pod_id,
             agent_id=conversation.agent_id,
@@ -334,6 +233,50 @@ class ConversationService(PauseResumeMixin):
 
         return await self.conversation_repository.update_conversation(conversation)
 
+    async def list_conversations(
+        self,
+        *,
+        pod_id: UUID,
+        agent_selection: ConversationAgentSelection[str],
+        user_id: UUID,
+        status: ConversationStatus | None = None,
+        type: ConversationType | None = None,
+        metadata_filters: dict[str, object] | None = None,
+        parent_id: UUID | None = None,
+        cursor: UUID | None = None,
+        limit: int = 20,
+    ) -> tuple[list[Conversation], UUID | None]:
+        """See `ConversationQueries.list_conversations`."""
+        return await self.queries.list_conversations(
+            pod_id=pod_id,
+            agent_selection=agent_selection,
+            user_id=user_id,
+            status=status,
+            type=type,
+            metadata_filters=metadata_filters,
+            parent_id=parent_id,
+            cursor=cursor,
+            limit=limit,
+        )
+
+    async def get_conversation(
+        self,
+        *,
+        conversation_id: UUID,
+        user_id: UUID,
+        pod_id: UUID,
+        agent_name: str | None = None,
+        require_read_grant: bool = True,
+    ) -> Conversation:
+        """See `ConversationQueries.get_conversation`."""
+        return await self.queries.get_conversation(
+            conversation_id=conversation_id,
+            user_id=user_id,
+            pod_id=pod_id,
+            agent_name=agent_name,
+            require_read_grant=require_read_grant,
+        )
+
     async def list_messages(
         self,
         *,
@@ -345,27 +288,12 @@ class ConversationService(PauseResumeMixin):
         after_sequence: int | None = None,
         limit: int = 100,
     ) -> tuple[list[Message], int | None]:
-        expected_agent_id = await self._expected_agent_id(
+        """See `ConversationQueries.list_messages`."""
+        return await self.queries.list_messages(
+            conversation_id=conversation_id,
+            user_id=user_id,
             pod_id=pod_id,
             agent_name=agent_name,
-        )
-        conversation = await self.conversation_repository.get_conversation(
-            conversation_id
-        )
-        validate_conversation_access(
-            conversation,
-            user_id=user_id,
-            pod_id=pod_id,
-            agent_id=expected_agent_id,
-        )
-        await self._require_agent_action(
-            user_id=user_id,
-            pod_id=pod_id,
-            agent_id=conversation.agent_id,
-            action=Permissions.AGENT_READ,
-        )
-        return await self.conversation_repository.list_messages(
-            conversation_id=conversation_id,
             before_sequence=before_sequence,
             after_sequence=after_sequence,
             limit=limit,
@@ -379,26 +307,13 @@ class ConversationService(PauseResumeMixin):
         pod_id: UUID,
         agent_name: str | None = None,
     ) -> AgentRun | None:
-        expected_agent_id = await self._expected_agent_id(
+        """See `ConversationQueries.get_active_agent_run`."""
+        return await self.queries.get_active_agent_run(
+            conversation_id=conversation_id,
+            user_id=user_id,
             pod_id=pod_id,
             agent_name=agent_name,
         )
-        conversation = await self.conversation_repository.get_conversation(
-            conversation_id
-        )
-        validate_conversation_access(
-            conversation,
-            user_id=user_id,
-            pod_id=pod_id,
-            agent_id=expected_agent_id,
-        )
-        await self._require_agent_action(
-            user_id=user_id,
-            pod_id=pod_id,
-            agent_id=conversation.agent_id,
-            action=Permissions.AGENT_READ,
-        )
-        return await self.conversation_repository.get_active_agent_run(conversation_id)
 
     async def list_user_approvals(
         self,
@@ -408,18 +323,13 @@ class ConversationService(PauseResumeMixin):
         pod_id: UUID,
         agent_name: str | None = None,
     ) -> list[Message]:
-        conversation = await self._authorized_conversation(
+        """See `ConversationQueries.list_user_approvals`."""
+        return await self.queries.list_user_approvals(
             conversation_id=conversation_id,
             user_id=user_id,
             pod_id=pod_id,
             agent_name=agent_name,
-            action=Permissions.AGENT_READ,
         )
-        messages, _ = await self.conversation_repository.list_messages(
-            conversation_id=conversation.id,
-            limit=500,
-        )
-        return pending_user_approval_messages(messages)
 
     async def resolve_user_approval(
         self,
@@ -446,7 +356,9 @@ class ConversationService(PauseResumeMixin):
         authorized the external user against the conversation owner) calls the
         internal method directly with the loaded conversation.
         """
-        conversation = await self._authorized_conversation(
+        conversation = await authorized_conversation(
+            self.conversation_repository,
+            self.agent_repository,
             conversation_id=conversation_id,
             user_id=user_id,
             pod_id=pod_id,
@@ -478,575 +390,19 @@ class ConversationService(PauseResumeMixin):
     ) -> ApprovalResolution:
         """Resume a paused run for an already-authorized + loaded conversation.
 
-        Idempotent and self-healing. Classifies the approval three ways:
-          * neither a paused call nor a recorded decision exists -> it is unknown
-            (``UnknownApprovalError`` -> 404);
-          * no decision yet -> record it (the unique (conversation, approval) row
-            is the double-submit lock), then reconcile;
-          * a decision already exists -> adopt it and reconcile (the retry /
-            self-heal path) instead of erroring.
-
-        The reconcile step is safe to call repeatedly: the approved tool runs at
-        most once (guarded by an existing tool return) and the resume run starts
-        at most once. Callers MUST have authorized the resolver against this
-        conversation first. The caller's current auth context must be the
-        conversation owner's, since an approved ``request_approval`` runs the
-        wrapped tool with that authority.
-
-        ``defer_reconciliation`` is for callers under a deadline (an HTTP
-        request, a platform webhook): the decision still commits here, but the
-        slow half is handed to a worker job and the status comes back
-        ``"queued"``.
+        Kept on the service because `agent_surfaces` reaches this and
+        `get_pending_user_interaction` off one import of this class; see
+        `ApprovalCoordinator` for what it does.
         """
-        decision_row = await self.conversation_repository.get_approval_decision(
-            conversation_id=conversation.id,
+        return await self.approvals.resolve_user_approval_internal(
+            conversation=conversation,
             approval_id=approval_id,
-        )
-        paused = await self._paused_call_from_messages(
-            conversation_id=conversation.id,
-            approval_id=approval_id,
-        )
-        if decision_row is None and paused is None:
-            # Nothing to resolve and nothing to heal — the approval never existed
-            # or its call was never persisted.
-            raise UnknownApprovalError()
-
-        if paused is None:
-            # Unreachable in practice: the pausing tool call message is never
-            # deleted, so a recorded decision always has its call. Guard anyway —
-            # we cannot faithfully rebuild a resume without the original call.
-            raise UnknownApprovalError()
-        kind = str(paused["kind"])
-        tool_args = paused["tool_args"] if isinstance(paused["tool_args"], dict) else {}
-        paused_run_id = paused["agent_run_id"]
-
-        if decision_row is None:
-            # Fresh resolve: record the decision. The unique (conversation,
-            # approval) row locks out a concurrent double-submit before any
-            # side-effecting tool runs.
-            decision_tool_name = (
-                "ask_user"
-                if kind == "ask_user"
-                else str(tool_args.get("tool_name") or "request_approval")
-            )
-            recorded = await self.conversation_repository.record_approval_decision(
-                conversation_id=conversation.id,
-                approval_id=approval_id,
-                agent_run_id=paused_run_id,
-                tool_name=decision_tool_name,
-                decision=decision,
-                response=response or {},
-                resolved_by_user_id=user_id,
-            )
-            await self.uow.commit()
-            if recorded:
-                effective_decision, effective_response, status = (
-                    decision,
-                    response or {},
-                    "resolved",
-                )
-            else:
-                # A concurrent resolve won the race; adopt its stored decision and
-                # reconcile (idempotent) rather than raising "already resolved".
-                stored = await self.conversation_repository.get_approval_decision(
-                    conversation_id=conversation.id,
-                    approval_id=approval_id,
-                )
-                effective_decision, effective_response = (
-                    stored if stored is not None else (decision, response or {})
-                )
-                status = "reconciled"
-        else:
-            # Decision already recorded (retry / self-heal). Do NOT re-record or
-            # re-run the tool; adopt the stored decision and finish whatever the
-            # prior attempt left undone.
-            effective_decision, effective_response = decision_row
-            status = "reconciled"
-
-        existing_return = (
-            await self.conversation_repository.get_tool_return(
-                conversation_id=conversation.id,
-                tool_call_id=approval_id,
-            )
-            if defer_reconciliation
-            else None
-        )
-        if should_defer_approved_tool(
+            user_id=user_id,
+            pod_id=pod_id,
+            decision=decision,
+            response=response,
+            agent_name=agent_name,
             defer_reconciliation=defer_reconciliation,
-            kind=kind,
-            tool_args=tool_args,
-            decision=effective_decision,
-            has_tool_return=existing_return is not None,
-        ):
-            # Deferred to after the commit, for two reasons. The connection is
-            # the smaller one: enqueuing is a Redis round trip and this runs
-            # inside the caller's transaction. The larger one is ordering -- an
-            # enqueue that happens before the commit queues a job against state
-            # that a rollback would erase, and the worker would then reconcile
-            # an approval the database never accepted.
-            self.uow.after_commit(
-                lambda: queue_approval_reconciliation(
-                    conversation_id=conversation.id,
-                    approval_id=approval_id,
-                    user_id=user_id,
-                    pod_id=pod_id,
-                )
-            )
-            return ApprovalResolution(status="queued", decision=effective_decision)
-
-        await self._reconcile_approval_resume(
-            conversation=conversation,
-            approval_id=approval_id,
-            paused_run_id=paused_run_id,
-            kind=kind,
-            tool_args=tool_args,
-            decision=effective_decision,
-            response=effective_response,
-            user_id=user_id,
-            pod_id=pod_id,
-            agent_name=agent_name,
-        )
-        return ApprovalResolution(status=status, decision=effective_decision)
-
-    async def _reconcile_approval_resume(
-        self,
-        *,
-        conversation: Conversation,
-        approval_id: str,
-        paused_run_id: UUID,
-        kind: str,
-        tool_args: dict[str, object],
-        decision: AgentRunApprovalDecision,
-        response: dict[str, object],
-        user_id: UUID,
-        pod_id: UUID,
-        agent_name: str | None,
-    ) -> None:
-        """Finish (or re-finish) an approval's resume; safe to call repeatedly.
-
-        Two idempotent halves:
-          1. Synthesize + persist the paused call's tool return exactly once. If a
-             return already exists the approved tool has already run, so we skip
-             the rebuild entirely — re-executing it (e.g. re-deploying an app, or
-             re-recording session grants) would be a correctness bug.
-          2. Start the resume run only once every pausing call in the paused run is
-             resolved and no run is already active. The conversation lock
-             serializes this so two near-simultaneous resolves don't each start a
-             run.
-
-        A resolve that died mid-flight (decision committed, return not appended,
-        or run not started) self-heals when this runs again on the user's retry.
-        """
-        existing_return = await self.conversation_repository.get_tool_return(
-            conversation_id=conversation.id,
-            tool_call_id=approval_id,
-        )
-        if existing_return is None:
-            # Build the return the resumed run will replay. This check guards the
-            # *build*, which for an approved request_approval runs the wrapped tool
-            # as the user — re-executing it (re-deploying an app, re-recording
-            # session grants) would be a correctness bug. The append below re-checks
-            # cheaply and closes the race.
-            return_tool_name, tool_result = await self._build_resume_tool_return(
-                conversation=conversation,
-                user_id=user_id,
-                kind=kind,
-                tool_args=tool_args,
-                decision=decision,
-                response=response,
-                paused_agent_run_id=paused_run_id,
-            )
-            await self.append_pause_tool_return(
-                conversation=conversation,
-                paused_run_id=paused_run_id,
-                tool_call_id=approval_id,
-                tool_name=return_tool_name,
-                tool_result=tool_result,
-            )
-
-        if agent_host_permission_request(tool_args) is not None:
-            # An Agent Host pauses *inside* a live run: the decision was just
-            # handed to the host, which carries the same run on from where it
-            # stopped. Starting a resume run here would dispatch a second,
-            # duplicate host run for the same turn.
-            return
-
-        if await self.resume_would_duplicate_a_live_turn(paused_run_id):
-            return
-
-        await self.start_resume_run_if_ready(
-            conversation=conversation,
-            paused_run_id=paused_run_id,
-            resumed_tool_call_id=approval_id,
-            user_id=user_id,
-            pod_id=pod_id,
-            agent_name=agent_name,
-            source="approval_resume",
-        )
-
-    async def _build_resume_tool_return(
-        self,
-        *,
-        conversation: Conversation,
-        user_id: UUID,
-        kind: str,
-        tool_args: dict[str, object],
-        decision: AgentRunApprovalDecision,
-        response: dict[str, object],
-        paused_agent_run_id: UUID,
-        deliver_to_host: bool = True,
-    ) -> tuple[str, object]:
-        """Return ``(tool_name, tool_result)`` for the synthesized resume message."""
-        from app.modules.agent.tools.user_interaction.models import (
-            AskUserResponse,
-            RequestApprovalResponse,
-        )
-
-        if kind == "ask_user":
-            if decision == AgentRunApprovalDecision.DENY:
-                content = AskUserResponse(
-                    success=False,
-                    message="User dismissed the questions without answering.",
-                )
-            else:
-                answers: dict[str, object] = {}
-                candidate = response.get("answers")
-                if isinstance(candidate, dict):
-                    answers = candidate
-                elif response:
-                    answers = response
-                content = AskUserResponse(
-                    success=True,
-                    answers=answers,
-                    message="User answered the questions.",
-                )
-            return "ask_user", content.model_dump(mode="json")
-
-        host_permission = agent_host_permission_request(tool_args)
-        if host_permission is not None and deliver_to_host:
-            # Checked before the denial branch below: a denial must reach the
-            # host too, or its ACP agent sits blocked until the request times
-            # out half an hour later.
-            return "request_approval", await agent_host_permission_tool_return(
-                uow=self.uow,
-                request=host_permission,
-                agent_run_id=paused_agent_run_id,
-                decision=decision,
-                response=response,
-            )
-        if host_permission is not None:
-            # Superseding rides the caller's uncommitted transaction. Handing
-            # the decision to the host from here would commit a command in a
-            # separate transaction that the caller's rollback could not take
-            # back. The run this belonged to is over, so there is nothing to
-            # unblock; a host still executing an orphaned run is stopped by
-            # reconcile_agent_host_dispatch, which cancels it outright.
-            return "request_approval", RequestApprovalResponse(
-                success=False,
-                message="The request was superseded before it was answered.",
-                decision=decision,
-                executed=False,
-                response=response,
-            ).model_dump(mode="json")
-
-        inner_tool = str(tool_args.get("tool_name") or "")
-        inner_args = tool_args.get("args")
-        inner_args = inner_args if isinstance(inner_args, dict) else {}
-        if decision == AgentRunApprovalDecision.DENY:
-            content = RequestApprovalResponse(
-                success=False,
-                message=f"User denied running {inner_tool}.",
-                decision=decision,
-                executed=False,
-                response=response,
-            )
-            return "request_approval", content.model_dump(mode="json")
-
-        if decision == AgentRunApprovalDecision.APPROVE_FOR_SESSION:
-            # Beyond the one-off run below, remember the approval so the
-            # workload can keep performing this action type in this
-            # conversation (the authorizer honors it as an ephemeral grant,
-            # which is the only unlock for DESTRUCTIVE_ACTIONS besides an
-            # explicit grant). The permission ids ride in the request_approval
-            # args, copied by the agent from the denied tool result.
-            # Queued, not awaited: a Redis write inline holds a connection
-            # inside an open write transaction, and a rollback must not leave an
-            # approval standing. Lands before the tool runs because
-            # `execute_approved_tool_as_user` commits first -- see
-            # `test_a_session_approval_is_recorded_before_the_tool_runs`.
-            self.uow.after_commit(
-                partial(
-                    record_session_approvals,
-                    conversation_id=conversation.id,
-                    agent_id=conversation.agent_id,
-                    tool_args=tool_args,
-                    user_id=user_id,
-                )
-            )
-
-        executed = await self._execute_approved_tool_as_user(
-            conversation=conversation,
-            user_id=user_id,
-            agent_run_id=paused_agent_run_id,
-            tool_name=inner_tool,
-            args=dict(inner_args),
-        )
-        if executed["ok"]:
-            content = RequestApprovalResponse(
-                success=True,
-                message=f"Approved; {inner_tool} executed as the user.",
-                decision=decision,
-                executed=True,
-                result=executed["value"],
-                response=response,
-            )
-        else:
-            content = RequestApprovalResponse(
-                success=False,
-                error=f"Approved, but running {inner_tool} failed: {executed['error']}",
-                decision=decision,
-                executed=False,
-                response=response,
-            )
-        return "request_approval", content.model_dump(mode="json")
-
-    async def _execute_approved_tool_as_user(
-        self,
-        *,
-        conversation: Conversation,
-        user_id: UUID,
-        agent_run_id: UUID,
-        tool_name: str,
-        args: dict[str, object],
-    ) -> dict[str, object]:
-        """Run an approved tool with the user's authority; never raise."""
-        deps = await self._build_resume_context(
-            conversation=conversation,
-            user_id=user_id,
-            agent_run_id=agent_run_id,
-        )
-        return await execute_approved_tool_as_user(
-            uow=self.uow,
-            deps=deps,
-            tool_name=tool_name,
-            args=args,
-        )
-
-    async def _build_resume_context(
-        self,
-        *,
-        conversation: Conversation,
-        user_id: UUID,
-        agent_run_id: UUID,
-    ):
-        """Rebuild the agent run context so an approved tool runs like in-run.
-
-        Mirrors ``AgentRunnerService.execute``'s context build (runtime profile,
-        workspace location, configured accounts). Surface delivery context is
-        omitted — approval-gated action tools don't deliver to surfaces.
-        """
-        from app.core.infrastructure.db.session import async_session_maker
-        from app.core.infrastructure.db.uow_factory import SessionUnitOfWorkFactory
-        from app.modules.agent.infrastructure.repositories import (
-            AgentRuntimeProfileRepository,
-        )
-        from app.modules.agent.services.runtime_profile_service import (
-            AgentRuntimeProfileService,
-        )
-        from app.modules.agent.tools.callable_tool_factory import (
-            AgentCallableToolFactory,
-        )
-        from app.modules.agent.tools.context import ConversationContext
-        from app.core.crypto import get_secret_cipher
-        from app.modules.agent.services.workspace_location import resolve_pod_cwd
-
-        uow_factory = SessionUnitOfWorkFactory(async_session_maker)
-        agent = await resolve_agent(
-            conversation,
-            user_id=user_id,
-            agent_repository=self.agent_repository,
-        )
-        selected_runtime = (
-            conversation.agent_runtime
-            or agent.agent_runtime
-            or await self._default_agent_runtime_for_pod(pod_id=conversation.pod_id)
-        )
-        async with uow_factory() as uow:
-            profile_service = AgentRuntimeProfileService(
-                AgentRuntimeProfileRepository(uow, encryption=get_secret_cipher())
-            )
-            resolved = await profile_service.resolve(
-                runtime=selected_runtime,
-                organization_id=conversation.organization_id,
-                user_id=user_id,
-            )
-        configured_accounts = await AgentCallableToolFactory(
-            uow_factory
-        ).resolve_configured_accounts(agent=agent, user_id=user_id)
-        workspace_location = resolve_workspace_location(conversation)
-        return ConversationContext(
-            user_id=user_id,
-            org_id=conversation.organization_id,
-            pod_id=conversation.pod_id,
-            conversation_id=conversation.id,
-            agent_name=agent.name,
-            agent_run_id=agent_run_id,
-            workload_type="agent",
-            workload_id=agent.id,
-            configured_accounts=configured_accounts,
-            runtime_profile=resolved.public_snapshot(),
-            runtime_credentials=resolved.credentials or {},
-            workspace_id=workspace_location.workspace_id,
-            workspace_cwd=workspace_location.cwd,
-            workspace_repo=workspace_location.repo,
-            pod_cwd=resolve_pod_cwd(conversation),
-        )
-
-    async def _supersede_stale_pending_interactions(
-        self,
-        *,
-        conversation: Conversation,
-        user_id: UUID,
-    ) -> list[Message]:
-        """Auto-deny any ask_user/request_approval call left unresolved from an
-        earlier WAITING pause, before starting a fresh run for a new message.
-
-        This only ever runs when no run is currently active (the caller checked
-        that already), so any pausing call still unresolved at this point belongs
-        to a run that already finished — the user moved on without answering it.
-        Always DENY, never approve: this is a safety fallback synthesizing a
-        response on the user's behalf, not a real decision, so a request_approval
-        must never auto-execute here. Writes ride the caller's transaction (no
-        commit here) so they land atomically with the new run/message it creates;
-        the caller is responsible for publishing the returned messages once that
-        transaction actually commits. That also rules out side effects Lemma
-        could not take back on a rollback, which is why an Agent Host permission
-        is not delivered from here.
-        """
-        resolved_ids = await self.conversation_repository.list_resolved_approval_ids(
-            conversation_id=conversation.id
-        )
-        messages, _ = await self.conversation_repository.list_messages(
-            conversation_id=conversation.id,
-            limit=500,
-        )
-        stale = [
-            message
-            for message in messages
-            if message.kind == MessageKind.TOOL_CALL
-            and message.tool_name in _PAUSING_TOOL_NAMES
-            and message.tool_call_id is not None
-            and message.tool_call_id not in resolved_ids
-        ]
-        synthesized_returns: list[Message] = []
-        for message in stale:
-            saved_return = await self._deny_stale_pending_interaction(
-                conversation=conversation,
-                message=message,
-                user_id=user_id,
-            )
-            if saved_return is not None:
-                synthesized_returns.append(saved_return)
-        return synthesized_returns
-
-    async def _deny_stale_pending_interaction(
-        self,
-        *,
-        conversation: Conversation,
-        message: Message,
-        user_id: UUID,
-    ) -> Message | None:
-        tool_args = message.tool_args if isinstance(message.tool_args, dict) else {}
-        decision_tool_name = (
-            "ask_user"
-            if message.tool_name == "ask_user"
-            else str(tool_args.get("tool_name") or "request_approval")
-        )
-        response = {"superseded_by_new_message": True}
-        recorded = await self.conversation_repository.record_approval_decision(
-            conversation_id=conversation.id,
-            approval_id=message.tool_call_id,
-            agent_run_id=message.agent_run_id,
-            tool_name=decision_tool_name,
-            decision=AgentRunApprovalDecision.DENY,
-            response=response,
-            resolved_by_user_id=user_id,
-        )
-        if not recorded:
-            # A concurrent resolve already recorded a real decision for this call;
-            # that resolve (or its own reconcile) owns synthesizing the return.
-            return None
-        existing_return = await self.conversation_repository.get_tool_return(
-            conversation_id=conversation.id,
-            tool_call_id=message.tool_call_id,
-        )
-        if existing_return is not None:
-            return None
-        return_tool_name, tool_result = await self._build_resume_tool_return(
-            conversation=conversation,
-            user_id=user_id,
-            kind=message.tool_name,
-            tool_args=tool_args,
-            decision=AgentRunApprovalDecision.DENY,
-            response=response,
-            paused_agent_run_id=message.agent_run_id,
-            deliver_to_host=False,
-        )
-        return await self.conversation_repository.append_message(
-            conversation_id=conversation.id,
-            agent_run_id=message.agent_run_id,
-            draft=MessageDraft.of_tool_return(
-                tool_call_id=message.tool_call_id,
-                tool_name=return_tool_name,
-                tool_result=tool_result,
-            ),
-        )
-
-    async def _paused_call_from_messages(
-        self,
-        *,
-        conversation_id: UUID,
-        approval_id: str,
-    ) -> dict[str, object] | None:
-        """The pausing tool call for an approval, regardless of decision state.
-
-        Addressed directly by ``tool_call_id`` (not a message-window scan) so a
-        long conversation can't hide the original call during resume
-        reconciliation. Returns ``{agent_run_id, kind, tool_args}`` or ``None``.
-        """
-        message = await self.conversation_repository.get_tool_call(
-            conversation_id=conversation_id,
-            tool_call_id=approval_id,
-        )
-        if (
-            message is None
-            or message.tool_name not in _PAUSING_TOOL_NAMES
-            or message.agent_run_id is None
-        ):
-            return None
-        tool_args = message.tool_args if isinstance(message.tool_args, dict) else {}
-        return {
-            "agent_run_id": message.agent_run_id,
-            "kind": message.tool_name,
-            "tool_args": tool_args,
-        }
-
-    async def _pending_user_approval_from_messages(
-        self,
-        *,
-        conversation_id: UUID,
-        approval_id: str,
-    ) -> dict[str, object] | None:
-        """The paused call only if it has NOT been resolved yet (else ``None``)."""
-        already_resolved = await self.conversation_repository.get_approval_decision(
-            conversation_id=conversation_id,
-            approval_id=approval_id,
-        )
-        if already_resolved is not None:
-            return None
-        return await self._paused_call_from_messages(
-            conversation_id=conversation_id,
-            approval_id=approval_id,
         )
 
     async def get_pending_ask_user(
@@ -1060,7 +416,7 @@ class ConversationService(PauseResumeMixin):
         ingress to render the questions on the surface (from a WAITING event) and
         to route a typed reply back into the run as the answer.
         """
-        return await self._oldest_unresolved_pause(
+        return await self.approvals.oldest_unresolved_pause(
             conversation_id=conversation_id,
             tool_names=("ask_user",),
         )
@@ -1076,40 +432,10 @@ class ConversationService(PauseResumeMixin):
         Returns ``{tool_call_id, kind, tool_args, agent_run_id}``. Used by
         surface ingress to route a typed reply back into the paused run.
         """
-        return await self._oldest_unresolved_pause(
+        return await self.approvals.oldest_unresolved_pause(
             conversation_id=conversation_id,
             tool_names=_PAUSING_TOOL_NAMES,
         )
-
-    async def _oldest_unresolved_pause(
-        self,
-        *,
-        conversation_id: UUID,
-        tool_names: Sequence[str],
-    ) -> dict[str, object] | None:
-        resolved_ids = await self.conversation_repository.list_resolved_approval_ids(
-            conversation_id=conversation_id
-        )
-        messages, _ = await self.conversation_repository.list_messages(
-            conversation_id=conversation_id,
-            limit=500,
-        )
-        for message in messages:
-            if (
-                message.kind == MessageKind.TOOL_CALL
-                and message.tool_name in tool_names
-                and message.tool_call_id is not None
-                and message.tool_call_id not in resolved_ids
-            ):
-                return {
-                    "tool_call_id": message.tool_call_id,
-                    "kind": message.tool_name,
-                    "tool_args": (
-                        message.tool_args if isinstance(message.tool_args, dict) else {}
-                    ),
-                    "agent_run_id": message.agent_run_id,
-                }
-        return None
 
     async def add_user_message_and_start_run(
         self,
@@ -1130,7 +456,8 @@ class ConversationService(PauseResumeMixin):
             require_grant=require_execute_grant,
         )
 
-        expected_agent_id = await self._expected_agent_id(
+        expected_agent_id = await resolve_expected_agent_id(
+            self.agent_repository,
             pod_id=pod_id,
             agent_name=agent_name,
         )
@@ -1144,205 +471,20 @@ class ConversationService(PauseResumeMixin):
         )
         # require_execute_grant is False only for self-spawn (SubAgentService).
         if require_execute_grant:
-            await self._require_agent_action(
+            await require_agent_action(
                 user_id=user_id,
                 pod_id=pod_id,
                 agent_id=conversation.agent_id,
                 action=Permissions.AGENT_EXECUTE,
             )
-        # Resolve the agent (a read) before taking the conversation lock, so the
-        # FOR UPDATE span covers only the active-run check + run/message writes.
-        agent = await resolve_agent(
+        return await self.turns.start(
             conversation,
             user_id=user_id,
-            agent_repository=self.agent_repository,
-        )
-
-        await self.conversation_repository.lock_conversation(conversation.id)
-        active_run = await self.conversation_repository.get_active_agent_run_for_update(
-            conversation.id
-        )
-        started_new_run = active_run is None
-        superseded_returns: list[Message] = []
-        if active_run is None:
-            # A prior run may have paused on ask_user/request_approval (conversation
-            # -> WAITING) without the user ever resolving it — the composer stays
-            # enabled during WAITING, so the user can type past the card. Deny any
-            # such leftover call now: otherwise this new run's history rebuild finds
-            # no matching return for it and silently drops it (see
-            # PydanticAIHarness._build_tool_batch), permanently losing the model's
-            # memory of asking and leaving the UI card stuck "needs your input".
-            superseded_returns = await self._supersede_stale_pending_interactions(
-                conversation=conversation,
-                user_id=user_id,
-            )
-            selected_agent_runtime = (
-                conversation.agent_runtime
-                or agent.agent_runtime
-                or await self._default_agent_runtime_for_pod(pod_id=conversation.pod_id)
-            )
-            await self._assert_usage_preflight_allowed(
-                organization_id=conversation.organization_id,
-                user_id=user_id,
-                agent_runtime=selected_agent_runtime,
-            )
-            active_run = await self.conversation_repository.create_agent_run(
-                conversation_id=conversation.id,
-                agent_id=conversation.agent_id,
-                agent_runtime=selected_agent_runtime,
-                metadata={"source": "user_message"},
-            )
-
-        metadata = {
-            "during_active_run": not started_new_run,
-            **(message_metadata or {}),
-        }
-        metadata.pop("author_user_id", None)
-        metadata.pop("agent_run_id", None)
-
-        saved_user_message = await self.conversation_repository.append_message(
-            conversation_id=conversation.id,
-            agent_run_id=active_run.id,
-            draft=MessageDraft.of_text(
-                content,
-                role=MessageRole.USER,
-                metadata=metadata,
-            ),
-        )
-
-        if started_new_run and not run_enqueue_suppressed():
-            self.uow.collect_events(
-                [
-                    AgentRunStartedEvent(
-                        conversation_id=conversation.id,
-                        agent_run_id=active_run.id,
-                        user_id=user_id,
-                        pod_id=pod_id,
-                        agent_name=agent_name,
-                    )
-                ]
-            )
-
-        # Streaming endpoints need the message/run and its outbox event committed
-        # atomically before the worker can safely load them; normal CRUD methods
-        # still rely on the request UoW.
-        await self.uow.commit()
-        # After the commit, not inside it: this claimed to run "now that they're
-        # durably committed", but the commit is the caller's, so it held a
-        # connection across a Redis round trip with the row locked. Not the
-        # outbox -- these are live UI frames; the next fetch recovers a lost one.
-        frames = [
-            message_payload(item.agent_run_id, message_to_payload(item))
-            for item in superseded_returns
-        ] + [input_added_payload(active_run.id, message_to_payload(saved_user_message))]
-
-        async def _publish_frames() -> None:
-            for frame in frames:
-                await publish_conversation_event(conversation.id, frame)
-
-        self.uow.after_commit(_publish_frames)
-        return AgentRunStartResult(
-            conversation_id=conversation.id,
-            agent_run_id=active_run.id,
-            started_new_run=started_new_run,
-        )
-
-    async def stop_conversation(
-        self,
-        *,
-        conversation_id: UUID,
-        user_id: UUID,
-        pod_id: UUID,
-        agent_name: str | None = None,
-    ) -> Conversation:
-        expected_agent_id = await self._expected_agent_id(
             pod_id=pod_id,
+            content=content,
             agent_name=agent_name,
+            message_metadata=message_metadata,
         )
-        conversation = await self.conversation_repository.get_conversation(
-            conversation_id
-        )
-        validate_conversation_access(
-            conversation,
-            user_id=user_id,
-            pod_id=pod_id,
-            agent_id=expected_agent_id,
-        )
-        await self._require_agent_action(
-            user_id=user_id,
-            pod_id=pod_id,
-            agent_id=conversation.agent_id,
-            action=Permissions.AGENT_EXECUTE,
-        )
-        active_run = await self.conversation_repository.get_active_agent_run_for_update(
-            conversation.id
-        )
-        if active_run is not None:
-            finish_result = await self.conversation_repository.finish_agent_run(
-                agent_run_id=active_run.id,
-                status=AgentRunStatus.STOP_REQUESTED,
-            )
-            if finish_result is not None:
-                conversation.status = finish_result.conversation_status
-            self.conversation_repository.collect_events(
-                [
-                    AgentRunStopRequestedEvent(
-                        conversation_id=conversation.id,
-                        agent_run_id=active_run.id,
-                        user_id=user_id,
-                    )
-                ]
-            )
-            await self.uow.commit()
-            return conversation
-
-        # No active run, but the conversation may still be suspended. A snoozed
-        # turn has *no* run by construction — it ended cleanly when the tool
-        # paused it — so without this, Stop silently did nothing and the timer
-        # still fired later.
-        await self._cancel_active_snooze(conversation=conversation)
-        return conversation
-
-    @property
-    def wait_repository(self) -> AgentConversationWaitRepository:
-        # Built on demand rather than in __init__: the repository binds a session
-        # eagerly, and plenty of callers construct this service without a real
-        # unit of work to exercise paths that never touch the database.
-        return AgentConversationWaitRepository(self.uow)
-
-    async def _cancel_active_snooze(self, *, conversation: Conversation) -> None:
-        """Stop a sleeping agent for good: drop the timer, never resume.
-
-        The CANCELLED tool return is still written, so the paused call is not
-        left dangling in history — a tool call with no return is dropped when
-        history is rebuilt, and the model would see a turn that ends mid-thought.
-        What is deliberately skipped is ``start_resume_run_if_ready``: Stop means
-        the agent does not wake.
-        """
-        wait = await self.wait_repository.find_active_for_conversation(conversation.id)
-        if wait is None:
-            return
-
-        wait.cancel()
-        await self.wait_repository.update(wait)
-        await self.conversation_repository.set_conversation_status(
-            conversation_id=conversation.id,
-            status=ConversationStatus.STOPPED,
-        )
-        conversation.status = ConversationStatus.STOPPED
-        await self.append_pause_tool_return(
-            conversation=conversation,
-            paused_run_id=wait.agent_run_id,
-            tool_call_id=wait.tool_call_id,
-            tool_name="snooze",
-            tool_result=build_snooze_result(
-                woke_because="CANCELLED",
-                slept_seconds=elapsed_seconds((wait.spec or {}).get("started_at")),
-                note_to_self=(wait.spec or {}).get("note_to_self"),
-            ),
-        )
-        if wait.external_ref:
-            await cancel_snooze_wake(wait.external_ref)
 
     async def _get_or_create_conversation_for_message(
         self,
@@ -1367,129 +509,26 @@ class ConversationService(PauseResumeMixin):
             user_id=user_id,
         )
 
-    async def _resolve_agent_for_path(
-        self,
-        *,
-        pod_id: UUID,
-        agent_name: str,
-    ) -> Agent:
-        agent = await self.agent_repository.get_by_pod_and_name(
-            pod_id=pod_id,
-            name=agent_name,
-        )
-        if agent is None:
-            raise AgentNotFoundError(agent_name)
-        return agent
-
-    async def _expected_agent_id(
-        self,
-        *,
-        pod_id: UUID,
-        agent_name: str | None,
-    ) -> UUID | None:
-        if agent_name is None:
-            return None
-        agent = await self._resolve_agent_for_path(
-            pod_id=pod_id,
-            agent_name=agent_name,
-        )
-        return agent.id
-
     async def _get_pod_organization_id(self, pod_id: UUID) -> UUID | None:
         return await create_agent_pod_repository(self.uow).get_organization_id(pod_id)
 
-    async def _default_agent_runtime_for_pod(
-        self,
-        *,
-        pod_id: UUID,
-    ) -> AgentRuntimeConfig:
-        config = await create_agent_pod_repository(self.uow).get_config(pod_id)
-        runtime = PodConfig.from_raw(config).resolved_default_runtime()
-        return runtime or AgentRuntimeConfig(
-            profile_id=DEFAULT_SYSTEM_AGENT_RUNTIME_PROFILE_ID
-        )
-
-    async def _assert_usage_preflight_allowed(
-        self,
-        *,
-        organization_id: UUID | None,
-        user_id: UUID,
-        agent_runtime: AgentRuntimeConfig,
-    ) -> None:
-        if self.usage_service is None:
-            return
-        if not agent_runtime.profile_id.startswith("system:"):
-            return
-        limits = await self.usage_service.get_usage_limits(
-            organization_id=organization_id,
-            user_id=user_id,
-        )
-        if limits["allowed"]:
-            return
-        raise UsageLimitExceededError()
-
-    async def _authorized_conversation(
+    async def stop_conversation(
         self,
         *,
         conversation_id: UUID,
         user_id: UUID,
         pod_id: UUID,
-        agent_name: str | None,
-        action: str,
+        agent_name: str | None = None,
     ) -> Conversation:
-        expected_agent_id = await self._expected_agent_id(
+        """Stop the run in flight, closing whatever pause it was sitting on."""
+        return await self.turns.stop_conversation(
+            conversation_id=conversation_id,
+            user_id=user_id,
             pod_id=pod_id,
             agent_name=agent_name,
         )
-        conversation = await self.conversation_repository.get_conversation(
-            conversation_id
-        )
-        validate_conversation_access(
-            conversation,
-            user_id=user_id,
-            pod_id=pod_id,
-            agent_id=expected_agent_id,
-        )
-        await self._require_agent_action(
-            user_id=user_id,
-            pod_id=pod_id,
-            agent_id=conversation.agent_id,
-            action=action,
-        )
-        return conversation
 
-    async def _require_agent_action(
-        self,
-        *,
-        user_id: UUID,
-        pod_id: UUID,
-        agent_id: UUID | None,
-        action: str,
-    ) -> None:
-        await self._require_agent_actions(
-            user_id=user_id,
-            pod_id=pod_id,
-            agent_id=agent_id,
-            actions=(action,),
-        )
-
-    async def _require_agent_actions(
-        self,
-        *,
-        user_id: UUID,
-        pod_id: UUID,
-        agent_id: UUID | None,
-        actions: Sequence[str],
-    ) -> None:
-        _ = user_id
-        ctx = get_current_context()
-        if ctx is None:
-            raise RuntimeError("Context is required for conversation authorization")
-        resource = ResourceRef(
-            resource_type=ResourceType.AGENT
-            if agent_id is not None
-            else ResourceType.POD,
-            resource_id=agent_id or pod_id,
-            pod_id=pod_id,
-        )
-        await ctx.require_all([(action, resource) for action in actions])
+    @property
+    def wait_repository(self) -> AgentConversationWaitRepository:
+        """The snooze timer store, reached through the turn coordinator."""
+        return self.turns.wait_repository
