+++ lemma-backend/app/modules/agent/services/conversation_turns.py
@@ -1,0 +1,318 @@
+"""Starting and stopping a turn.
+
+A turn is one user message plus the run that answers it. Starting one is mostly
+a question of *whether* to start a run at all -- a message typed while a run is
+already going joins that run instead of racing a second one -- and the answer
+has to be decided under the conversation lock, because two browser tabs asking
+at once must not each get a run.
+
+Stopping is the same question in reverse, and the asymmetry is deliberate: a
+stop closes any pause the run was sitting on, but never starts the resume run
+that closing a pause normally triggers. Stop means stop.
+
+Split from `ConversationService` because this is the part with the lock, the
+ordering constraints and the outbox; the rest of that class is storage.
+"""
+
+from __future__ import annotations
+
+from uuid import UUID
+
+from app.core.authorization.permissions import Permissions
+from app.core.infrastructure.db.uow import SqlAlchemyUnitOfWork
+from app.core.log.log import get_logger
+from app.composition.agent_snooze_scheduler import cancel_snooze_wake
+from app.composition.agent_usage import UsageLimitExceededError, UsageService
+from app.modules.agent.domain.entities import Conversation, Message
+from app.modules.agent.domain.events import (
+    AgentRunStartedEvent,
+    AgentRunStopRequestedEvent,
+)
+from app.modules.agent.domain.value_objects import (
+    AgentRunStartResult,
+    AgentRunStatus,
+    AgentRuntimeConfig,
+    ConversationStatus,
+    MessageDraft,
+    MessageRole,
+)
+from app.modules.agent.infrastructure.wait_repository import (
+    AgentConversationWaitRepository,
+)
+from app.modules.agent.services.conversation_access import (
+    require_agent_action,
+    resolve_agent,
+    resolve_expected_agent_id,
+    validate_conversation_access,
+)
+from app.modules.agent.services.conversation_approvals import ApprovalCoordinator
+from app.modules.agent.services.pause_resume import PauseResume
+from app.modules.agent.services.pod_runtime_defaults import (
+    default_agent_runtime_for_pod,
+)
+from app.modules.agent.services.realtime import (
+    input_added_payload,
+    message_payload,
+    publish_conversation_event,
+)
+from app.modules.agent.services.run_dispatch import run_enqueue_suppressed
+from app.modules.agent.services.serialization import message_to_payload
+from app.modules.agent.tools.snooze.models import (
+    build_snooze_result,
+    elapsed_seconds,
+)
+
+logger = get_logger(__name__)
+
+
+class TurnCoordinator:
+    """Starts the run that answers a message, and stops the one in flight."""
+
+    def __init__(
+        self,
+        uow: SqlAlchemyUnitOfWork,
+        conversation_repository: object,
+        agent_repository: object,
+        approvals: ApprovalCoordinator,
+        pauses: PauseResume,
+        usage_service: UsageService | None,
+    ) -> None:
+        self.uow = uow
+        self.conversation_repository = conversation_repository
+        self.agent_repository = agent_repository
+        self.approvals = approvals
+        self.pauses = pauses
+        self.usage_service = usage_service
+
+    async def start(
+        self,
+        conversation: Conversation,
+        *,
+        user_id: UUID,
+        pod_id: UUID,
+        content: str,
+        agent_name: str | None,
+        message_metadata: dict[str, object] | None,
+    ) -> AgentRunStartResult:
+        """Append the message and return the run that will answer it.
+
+        The conversation is already loaded and access-checked by the caller --
+        this is the half that needs the lock.
+        """
+        # Resolve the agent (a read) before taking the conversation lock, so the
+        # FOR UPDATE span covers only the active-run check + run/message writes.
+        agent = await resolve_agent(
+            conversation,
+            user_id=user_id,
+            agent_repository=self.agent_repository,
+        )
+
+        await self.conversation_repository.lock_conversation(conversation.id)
+        active_run = await self.conversation_repository.get_active_agent_run_for_update(
+            conversation.id
+        )
+        started_new_run = active_run is None
+        superseded_returns: list[Message] = []
+        if active_run is None:
+            # A prior run may have paused on ask_user/request_approval (conversation
+            # -> WAITING) without the user ever resolving it — the composer stays
+            # enabled during WAITING, so the user can type past the card. Deny any
+            # such leftover call now: otherwise this new run's history rebuild finds
+            # no matching return for it and silently drops it (see
+            # PydanticAIHarness._build_tool_batch), permanently losing the model's
+            # memory of asking and leaving the UI card stuck "needs your input".
+            superseded_returns = (
+                await self.approvals.supersede_stale_pending_interactions(
+                    conversation=conversation,
+                    user_id=user_id,
+                )
+            )
+            selected_agent_runtime = (
+                conversation.agent_runtime
+                or agent.agent_runtime
+                or await default_agent_runtime_for_pod(
+                    self.uow, pod_id=conversation.pod_id
+                )
+            )
+            await self._assert_usage_preflight_allowed(
+                organization_id=conversation.organization_id,
+                user_id=user_id,
+                agent_runtime=selected_agent_runtime,
+            )
+            active_run = await self.conversation_repository.create_agent_run(
+                conversation_id=conversation.id,
+                agent_id=conversation.agent_id,
+                agent_runtime=selected_agent_runtime,
+                metadata={"source": "user_message"},
+            )
+
+        metadata = {
+            "during_active_run": not started_new_run,
+            **(message_metadata or {}),
+        }
+        metadata.pop("author_user_id", None)
+        metadata.pop("agent_run_id", None)
+
+        saved_user_message = await self.conversation_repository.append_message(
+            conversation_id=conversation.id,
+            agent_run_id=active_run.id,
+            draft=MessageDraft.of_text(
+                content,
+                role=MessageRole.USER,
+                metadata=metadata,
+            ),
+        )
+
+        if started_new_run and not run_enqueue_suppressed():
+            self.uow.collect_events(
+                [
+                    AgentRunStartedEvent(
+                        conversation_id=conversation.id,
+                        agent_run_id=active_run.id,
+                        user_id=user_id,
+                        pod_id=pod_id,
+                        agent_name=agent_name,
+                    )
+                ]
+            )
+
+        # Streaming endpoints need the message/run and its outbox event committed
+        # atomically before the worker can safely load them; normal CRUD methods
+        # still rely on the request UoW.
+        await self.uow.commit()
+        # After the commit, not inside it: this claimed to run "now that they're
+        # durably committed", but the commit is the caller's, so it held a
+        # connection across a Redis round trip with the row locked. Not the
+        # outbox -- these are live UI frames; the next fetch recovers a lost one.
+        frames = [
+            message_payload(item.agent_run_id, message_to_payload(item))
+            for item in superseded_returns
+        ] + [input_added_payload(active_run.id, message_to_payload(saved_user_message))]
+
+        async def _publish_frames() -> None:
+            for frame in frames:
+                await publish_conversation_event(conversation.id, frame)
+
+        self.uow.after_commit(_publish_frames)
+        return AgentRunStartResult(
+            conversation_id=conversation.id,
+            agent_run_id=active_run.id,
+            started_new_run=started_new_run,
+        )
+
+    async def stop_conversation(
+        self,
+        *,
+        conversation_id: UUID,
+        user_id: UUID,
+        pod_id: UUID,
+        agent_name: str | None = None,
+    ) -> Conversation:
+        expected_agent_id = await resolve_expected_agent_id(
+            self.agent_repository,
+            pod_id=pod_id,
+            agent_name=agent_name,
+        )
+        conversation = await self.conversation_repository.get_conversation(
+            conversation_id
+        )
+        validate_conversation_access(
+            conversation,
+            user_id=user_id,
+            pod_id=pod_id,
+            agent_id=expected_agent_id,
+        )
+        await require_agent_action(
+            user_id=user_id,
+            pod_id=pod_id,
+            agent_id=conversation.agent_id,
+            action=Permissions.AGENT_EXECUTE,
+        )
+        active_run = await self.conversation_repository.get_active_agent_run_for_update(
+            conversation.id
+        )
+        if active_run is not None:
+            finish_result = await self.conversation_repository.finish_agent_run(
+                agent_run_id=active_run.id,
+                status=AgentRunStatus.STOP_REQUESTED,
+            )
+            if finish_result is not None:
+                conversation.status = finish_result.conversation_status
+            self.conversation_repository.collect_events(
+                [
+                    AgentRunStopRequestedEvent(
+                        conversation_id=conversation.id,
+                        agent_run_id=active_run.id,
+                        user_id=user_id,
+                    )
+                ]
+            )
+            await self.uow.commit()
+            return conversation
+
+        # No active run, but the conversation may still be suspended. A snoozed
+        # turn has *no* run by construction — it ended cleanly when the tool
+        # paused it — so without this, Stop silently did nothing and the timer
+        # still fired later.
+        await self._cancel_active_snooze(conversation=conversation)
+        return conversation
+
+    @property
+    def wait_repository(self) -> AgentConversationWaitRepository:
+        # Built on demand rather than in __init__: the repository binds a session
+        # eagerly, and plenty of callers construct this service without a real
+        # unit of work to exercise paths that never touch the database.
+        return AgentConversationWaitRepository(self.uow)
+
+    async def _cancel_active_snooze(self, *, conversation: Conversation) -> None:
+        """Stop a sleeping agent for good: drop the timer, never resume.
+
+        The CANCELLED tool return is still written, so the paused call is not
+        left dangling in history — a tool call with no return is dropped when
+        history is rebuilt, and the model would see a turn that ends mid-thought.
+        What is deliberately skipped is ``start_resume_run_if_ready``: Stop means
+        the agent does not wake.
+        """
+        wait = await self.wait_repository.find_active_for_conversation(conversation.id)
+        if wait is None:
+            return
+
+        wait.cancel()
+        await self.wait_repository.update(wait)
+        await self.conversation_repository.set_conversation_status(
+            conversation_id=conversation.id,
+            status=ConversationStatus.STOPPED,
+        )
+        conversation.status = ConversationStatus.STOPPED
+        await self.pauses.append_pause_tool_return(
+            conversation=conversation,
+            paused_run_id=wait.agent_run_id,
+            tool_call_id=wait.tool_call_id,
+            tool_name="snooze",
+            tool_result=build_snooze_result(
+                woke_because="CANCELLED",
+                slept_seconds=elapsed_seconds((wait.spec or {}).get("started_at")),
+                note_to_self=(wait.spec or {}).get("note_to_self"),
+            ),
+        )
+        if wait.external_ref:
+            await cancel_snooze_wake(wait.external_ref)
+
+    async def _assert_usage_preflight_allowed(
+        self,
+        *,
+        organization_id: UUID | None,
+        user_id: UUID,
+        agent_runtime: AgentRuntimeConfig,
+    ) -> None:
+        if self.usage_service is None:
+            return
+        if not agent_runtime.profile_id.startswith("system:"):
+            return
+        limits = await self.usage_service.get_usage_limits(
+            organization_id=organization_id,
+            user_id=user_id,
+        )
+        if limits["allowed"]:
+            return
+        raise UsageLimitExceededError()
